// frontend/src/api/generated/zod/transaction-templates/transaction-templates.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import { z as zod } from "zod";

export const getTemplateByIdParams = zod.object({
  id: zod.string().uuid(),
});

export const updateTemplateParams = zod.object({
  id: zod.string().uuid(),
});

export const updateTemplateBodyAmountMin = 0.01;

export const updateTemplateBody = zod.object({
  name: zod.string(),
  amount: zod.number().min(updateTemplateBodyAmountMin).optional(),
  description: zod.string().optional(),
  categoryId: zod.string().uuid(),
});

export const deleteTemplateParams = zod.object({
  id: zod.string().uuid(),
});

export const createTemplateBodyAmountMin = 0.01;

export const createTemplateBody = zod.object({
  name: zod.string(),
  amount: zod.number().min(createTemplateBodyAmountMin).optional(),
  description: zod.string().optional(),
  categoryId: zod.string().uuid(),
});